lib/db: Upgrade blobloom dep to build on 32-bit systems

+++ lib/db/lowlevel.go
@@ -27,7 +27,7 @@ const (
 	// than 100k. For fewer than 100k items we will just get better false
 	// positive rate instead.
 	indirectGCBloomCapacity          = 100000
-	indirectGCBloomFalsePositiveRate = 0.01    // 1%
+	indirectGCBloomFalsePositiveRate = 0.01     // 1%
 	indirectGCBloomMaxMem            = 32 << 20 // Use at most 32MiB memory, which covers our desired FP rate at 27 M items
 	indirectGCDefaultInterval        = 13 * time.Hour
 	indirectGCTimeKey                = "lastIndirectGCTime"
@@ -599,7 +599,7 @@ func (db *Lowlevel) gcIndirect(ctx context.Context) error {
 	blockFilter := blobloom.NewOptimized(blobloom.Config{
 		FPRate:  indirectGCBloomFalsePositiveRate,
 		MaxBits: 8 * indirectGCBloomMaxMem,
-		NKeys:   capacity,
+		NKeys:   uint64(capacity),
 	})
 
 	// Iterate the FileInfos, unmarshal the block and version hashes and
